config data 12

+++ services/insertMarks/src/main/java/com/stage/insertMarks/dto/UserDto.java
@@ -10,7 +10,7 @@ public record UserDto(
         String firstName,
         String lastName,
         String email,
-        Long idGrade,
+        Integer idGrade,
         String role
 ) {
 }
